I think i sucessfully replaced the vehicle related variables inside @code and moved AddVehicle entirely into bookingprocessor

+++ CarRental.Data/Classes/CollectionData.cs
@@ -56,7 +56,7 @@ public class CollectionData : IData
         {
             if(vehicle.VehicleStatus == VehicleStatuses.Available)
             {
-                int initialOdometer = vehicle.Odometer;
+                int initialOdometer = (int)vehicle.Odometer;
 
                 IBooking booking = new Booking(vehicle, customer, initialOdometer, DateTime.Now, VehicleBookingStatuses.Open);
